# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIGEA REST API",
        "description": "API REST del Sistema Integral de Gestión Educativa Académica (SIGEA). Autenticación mediante tokens Sanctum (header `Authorization: Bearer <token>`).",
        "contact": {
            "name": "Equipo SIGEA",
            "email": "contacto@sigea.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/perfil": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Perfil del usuario autenticado (con alumno/docente asociado)",
                "operationId": "d139941c8a07af6bec88f3d01e5139a6",
                "responses": {
                    "200": {
                        "description": "Datos del perfil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Actualizar nombre o email del usuario",
                "operationId": "79e3f1bb372b1ab414f06ad797d6404e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@sigea.local"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/perfil/password": {
            "put": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Cambiar la contraseña",
                "operationId": "b29d7df647456c412d658a6353080d66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada"
                    },
                    "422": {
                        "description": "Contraseña actual incorrecta o no coincide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}